Move sub components to right directory
